backend/requirements.txt
Django==4.2.7
Pillow==11.1.0
django-crispy-forms==2.1
django-widget-tweaks==1.5.0
google-generativeai==0.3.1
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
reportlab==4.0.7
channels==4.0.0
daphne==4.0.0
sendgrid==6.11.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
